Added travis file.

+++ avara-engine/src/Box.cc
@@ -50,7 +50,7 @@ Box::Box(float width, float height, float length):
 		
 		auto faces = vector<Face>();
 		for (const Triangle& t : box.triangles()) {
-			Face face = { (unsigned int)t.vertices[0], (unsigned int)t.vertices[1], (unsigned int)t.vertices[2] };
+			Face face = { t.vertices[0], t.vertices[1], t.vertices[2] };
 			faces.push_back(face);
 		}
