+++ avara-engine/src/Capsule.cc
@@ -62,7 +62,7 @@ Capsule::Capsule(float radius, float height, int slices, int segments, int rings
 			
 		auto faces = vector<Face>();
 		for (const Triangle& t : capsule.triangles()) {
-			Face face = { (unsigned int)t.vertices[0], (unsigned int)t.vertices[1], (unsigned int)t.vertices[2] };
+			Face face = { t.vertices[0], t.vertices[1], t.vertices[2] };
 			faces.push_back(face);
 		}
