+++ avara-engine/src/Torus.cc
@@ -60,7 +60,7 @@ Torus::Torus(float minorRadius, float majorRadius, int slices, int segments):
 		
 		auto faces = vector<Face>();
 		for (const Triangle& t : torus.triangles()) {
-			Face face = { (unsigned int)t.vertices[0], (unsigned int)t.vertices[1], (unsigned int)t.vertices[2] };
+			Face face = { t.vertices[0], t.vertices[1], t.vertices[2] };
 			faces.push_back(face);
 		}
