+++ avara-engine/src/Tube.cc
@@ -59,7 +59,7 @@ Tube::Tube(float innerRadius, float outerRadius, float height, int slices, int s
 		
 		auto faces = vector<Face>();
 		for (const Triangle& t : tube.triangles()) {
-			Face face = { (unsigned int)t.vertices[0], (unsigned int)t.vertices[1], (unsigned int)t.vertices[2] };
+			Face face = { t.vertices[0], t.vertices[1], t.vertices[2] };
 			faces.push_back(face);
 		}
